+++ core/src/com/nithinmuthukumar/conquest/Components/WeaponComponent.java
@@ -2,12 +2,13 @@ package com.nithinmuthukumar.conquest.Components;
 
 //identifies an entity that can do damage and how much damage it does
 public class WeaponComponent implements BaseComponent {
-    public int damage;
+    public float damage;
     @Override
     public BaseComponent create() {
         return this;
     }
-    public WeaponComponent create(int damage){
+
+    public WeaponComponent create(float damage) {
         this.damage=damage;
         return this;
     }
